backend/requirements.txt
# ── Web 框架 ──
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── 数据库 ──
asyncpg>=0.29.0                 # PG async driver
psycopg[binary,pool]>=3.1.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
neo4j>=5.14.0                   # Neo4j driver
pymilvus>=2.4.0                 # Milvus client

# ── Redis / 异步任务 ──
redis[hiredis]>=5.0.0
celery>=5.3.0
celery[redis]>=5.3.0

# ── AI / LLM ──
openai>=1.10.0                  # DeepSeek 兼容 OpenAI SDK
sentence-transformers>=2.2.0    # 本地 embedding
torch>=2.0.0                    # sentence-transformers 依赖

# ── 多模态处理 ──
openai-whisper>=20231106        # ASR (可选，可换 faster-whisper)
pytesseract>=0.3.10             # OCR
Pillow>=10.0.0
pdfplumber>=0.10.0              # PDF 文本提取
openpyxl>=3.1.0                 # Excel
python-docx>=1.0.0              # Word

# ── 文本处理 ──
tiktoken>=0.5.0                 # token 分块
spacy>=3.7.0                    # NER (英文)
jieba>=0.42.0                   # 中文分词

# ── 工具库 ──
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=23.0.0
uuid>=1.30
pyyaml>=6.0.1
tenacity>=8.2.0                 # retry
structlog>=23.0.0               # 结构化日志
